Set `CUDA_HOME` to `BUILD_PREFIX` if undefined
Simplify check for `CUDA_HOME` to merely define it if it undefined. This
would only happen for CUDA 12.0+ as `CUDA_HOME` was defined for CUDA
11.x and earlier. In the CUDA 12.0+ case, all of the CTK is installed as
Conda packages. For NCCL specifically, what we need is in
`BUILD_PREFIX` (namely the `cudart` static library). So point
`CUDA_HOME` to `BUILD_PREFIX` for CUDA 12.0+.